found circle bounds from centeral lat, long and radius

+++ map-service/src/main/java/com/example/mapservice/model/EventEntity.java
@@ -17,8 +17,8 @@ import lombok.Data;
 public class EventEntity extends BaseEntity{
     
     @Column(name = "latitude")
-    private Integer latitude;
+    private Double latitude;
     
     @Column(name = "longitude")
-    private Integer longitude;
+    private Double longitude;
 }
